add idmun in user

+++ Backend/src/main/java/com/example/mini_projet/Dto/Request/RegisterRequest.java
@@ -53,7 +53,7 @@ public class RegisterRequest {
     private String password;
 
     //@NotBlank
-    private Long codeMun;
+    private Integer codeMun;
