+++ Backend/src/main/java/com/example/mini_projet/User/UserService.java
@@ -40,7 +40,7 @@ public class UserService {
         throw new IllegalStateException("User not found");
     }
 
-    public List<User> getUsersByCodeMun(@PathVariable Long codeMun){
+    public List<User> getUsersByCodeMun(@PathVariable Integer codeMun){
         List<User> users = userRepository.findAll();
         List<User> usersByMun =
                 users.stream().filter(usr -> usr.getMunicipality().equals(municipalityRepository.findById(codeMun).get()))
